agent/requirements.txt
marshmallow==3.20.2
packaging==23.2
paho-mqtt==1.6.1
